requirements.txt
# Core
ccxt>=4.2.0
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0          # Parquet I/O

# ML
lightgbm>=4.2.0
scikit-learn>=1.3.0
optuna>=3.5.0             # Hyperparameter tuning

# Technical Analysis
ta>=0.11.0                # Technical indicators
pandas-ta>=0.3.14b        # Extended TA library

# Backtesting
vectorbt>=0.26.0          # Vectorized backtesting

# Bybit / Trading
pybit>=5.6.0              # Bybit SDK v5

# Dashboard
streamlit>=1.30.0
plotly>=5.18.0
streamlit-autorefresh>=1.0.1

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
loguru>=0.7.0
rich>=13.7.0              # Pretty CLI output
schedule>=1.2.0

# Dev
pytest>=7.4.0
black>=23.0.0
